don't ever show a scroll bar in the autocompleter drop-down...instead, show only 4 entries and then underneath them a hint saying "see all ___ matching friends" which pops open the sidebar. there were lots of bugs with the scrollbar anyway (which i blame firefox for), and this looks nicer and exposes the sidebar a bit better. also just cleaned up all other the "hinting" in the autocompleter so that the search code is quite a bit cleaner and so that instead of constantly adding/removing elements from the document we just always go with the "display: none;" toggling trick.
also automatically resize the status messages in the sidebar so that they fit
when you resize the sidebar.

also display ?'s for the poke/msg/req counts when you are not logged in.

i think there was some other stuff in here also but that was the bulk of it.