toolbar - fix retrying api calls
Summary: it wasn't re-using the correct parameters when retrying an api call
(after an unexpected failure).  thus, it would always wind up failing twice,
which was kind of useless.  now we make a deep copy of the parameters so that
the retry has a chance of succeeding.

Reviewed By: eugene

Test Plan: look carefully at code

Revert Plan: revertible